Crate ascii85

Implements the ascii85 data encoding as used in the btoa tool and Adobe’s PostScript and PDF document formats.

Structs

CorruptInputError

Error occurs during decoding.

Functions

decode

Decodes src into dst, returning both the number of bytes written to dst and the number consumed from src. If src contains invalid ascii85 data, decode will return a CorruptInputError. decode ignores space and control characters in src. Often, ascii85-encoded data is wrapped in <~ and ~> symbols. decode expects these to have been stripped by the caller.

encode

Encodes src into at most max_encoded_len(src.len()) bytes of dst, returning the actual number of bytes written. The encoding handles 4-byte chunks, using a special encoding for the last fragment, so encode is not appropriate for use on individual blocks of a large data stream. Use new_encoder() instead. Often, ascii85-encoded data is wrapped in <~ and ~> symbols. encode does not add these.

max_encoded_len

Returns the maximum length of an encoding of n source bytes.

new_decoder

Constructs a new ascii85 stream decoder.

new_encoder

Returns a new ascii85 stream encoder. Data written to the returned writer will be encoded and then written to w. Ascii85 encodings operate in 32-bit blocks.